Changing recurrence to count from tickle date
When an item is recurring, and it has a tickle date, that should be the
base for the recurrence. A good example are tasks which I'd like to do
anywhere from the 1st to the Nth of the month, like paying bills. These
should always tickle on the 1st of the next month and not shift depending
on the completion date.